+++ src/main/java/onboarding/Problem1/Pages.java
@@ -1,4 +1,4 @@
-package onboarding;
+package onboarding.Problem1;
 
 import java.util.List;
 
@@ -85,23 +85,4 @@ class Pages implements Comparable<Pages>
         else return DRAW_VALUE;
 
     }
-}
-
-public class Problem1 {
-
-    private static final int EXCEPTION_VALUE = -1;
-
-    public static int solution(List<Integer> pobi, List<Integer> crong) {
-        try {
-            Pages pobiPage = new Pages(pobi);
-            Pages crongPage = new Pages(crong);
-
-            return pobiPage.compareTo(crongPage);
-        } catch (IllegalArgumentException e) {
-            return EXCEPTION_VALUE;
-        }
-
-
-    }
-
 }
