+++ src/main/java/onboarding/Problem1/Problem1.java
@@ -1,0 +1,22 @@
+package onboarding.Problem1;
+
+import java.util.List;
+
+public class Problem1 {
+
+    private static final int EXCEPTION_VALUE = -1;
+
+    public static int solution(List<Integer> pobi, List<Integer> crong) {
+        try {
+            Pages pobiPage = new Pages(pobi);
+            Pages crongPage = new Pages(crong);
+
+            return pobiPage.compareTo(crongPage);
+        } catch (IllegalArgumentException e) {
+            return EXCEPTION_VALUE;
+        }
+
+
+    }
+
+}
